BREAK: move `simplify_latex_rendering()` to `io` module
