+++ src/ampform_dpd/io.py
@@ -378,3 +378,14 @@ def _warn_about_unsafe_hash():
     """
     message = dedent(message).replace("\n", " ").strip()
     _LOGGER.warning(message)
+
+
+def simplify_latex_rendering() -> None:
+    """Improve LaTeX rendering of an `~sympy.tensor.indexed.Indexed` object."""
+
+    def _print_Indexed_latex(self, printer, *args):  # noqa: N802
+        base = printer._print(self.base)
+        indices = ", ".join(map(printer._print, self.indices))
+        return f"{base}_{{{indices}}}"
+
+    sp.Indexed._latex = _print_Indexed_latex
